requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0